Changed UNIT so that it can store double values from now on

+++ Unit.cpp
@@ -1,11 +1,11 @@
 #include "Unit.h"
 
-int Unit::getHp() const
+double Unit::getHp() const
 {
 	return hp;
 }
 
-int Unit::getDmg() const
+double Unit::getDmg() const
 {
 	return dmg;
 }
@@ -17,7 +17,7 @@ std::string Unit::getName() const
 
 void Unit::takeDamage(const Unit& enemy)
 {
-	int damage = enemy.getDmg();
+	double damage = enemy.getDmg();
 	hp -= damage;
 	if (hp < 0)
 	{
@@ -36,6 +36,6 @@ Unit* Unit::parseUnit(const std::string& data){
 	std::cout << returnedMap["hp"] << std::endl;
 	std::cout << returnedMap["dmg"] << std::endl;
 	std::cout << returnedMap["name"] << std::endl;
-	Unit* hi = new Unit(std::stoi(returnedMap["hp"]),std::stoi(returnedMap["dmg"]),returnedMap["name"]);
+	Unit* hi = new Unit(std::stod(returnedMap["hp"]),std::stod(returnedMap["dmg"]),returnedMap["name"]);
 	return hi;
 }
